initial procedure implementation
Procedures will be "functions" that can have side effects. This
initial implementation allows a simple procedure without any
parameters to be defined, called and returned from. It required
adding call and ret instruction support, a return keyword, and an
adjustable entrypoint for the program.

For the sake of simplicity, it would probably be a good idea to have
a main function as an entrypoint, so toplevel items can only be
declarations.