Add types, stack frame construction and consequently procedures
Types are stored in a hashmap, indexed by their blake3 hash, with
an offset as an array as their value. This basically allows us to
ignore the possibility of hash collisions, because blake3 has aribtrary
width.

Type names are stored in a hashmap, indexed by their name, and have
an offset in an array as their value.

This makes it easy to find redundant types, even if they have the
same name. If a new type isn't in the first hashmap, then it is
added and added to the types array. Then an entry in the second
hashmap is created to point to the offset.

This was all done to allow parameter type specification for procedures,
although we don't typecheck them yet.

%rbp and %rsp are used for dealing with stack frames, the same way as
in C. Some work needs to be done to make sure that these are not
overwritten.

This enables actual procedures, so some tests for procedures for
the exit and write syscalls were added.

The map implementation is stolen from cproc (Michael Forney), and
blake3 comes from blake3-tiny (Michael Forney).