gensyscall: correct register allocation
Previously, we ignored if a syscall arg register was currently
allocated. This is obviously bad, because we replace the value
without saving the old value. This could cause problems with nested
syscalls, and more generally, nested function calls (once we pass
parameters in registers). Now we push the value of used registers
to the stack, perform the syscall, and then pop the values back off
the stack.