return 1 value from a procedure
Instead of genexpr taking a register as output, it now takes a
struct out, which let's us abstract over storing to a register and
memory. The address to store the output to (a decl's address) is
given in the struct out when we do an fcall, otherwise a register
is given, because that's what we do for all other expression types.

This makes code generation much less efficient: unnecessary registers
are allocated all over the place, and essentially 2 movs each for
what should be 1 mov.

Instructions were added which store and load from memory whose address
is specified in a register.