add x64 backend and do basic register allocation
A lot of cleanup can be done, but all tests pass.

The register allocation algorithm is linear scan, but implemented
naively.

To allow for multiple returns in the future, the IR has no concept
of a return value. Instead we pass pointers to where to put the
return values after all the normal parameters. Typechecking had to
be adjusted to account for this, although perhaps there's a nicer
way where typechecking can be uniform.