ir: add iblock for more correct register allocation
The old register allocation method didn't take into account blocks,
so inside a conditional loop for example, a register would be freed
in the middle of the loop's block and reused in the condition, even
though the value was still needed.

This fixes that because we extend a temporary's allocation to the
end of any sub-blocks where it is used. Perhaps in the future, we
can have some sort of flag to indicate that temporaries used in a
block shouldn't be freed, which will allow us to free temporaries
early in non-repeating blocks.